fix: dedicated module for previews
