+++ style-system/src/commonMain/kotlin/io/timemates/app/style/system/appbar/AppBar.kt
@@ -1,16 +1,7 @@
 package io.timemates.app.style.system.appbar
 
-import androidx.compose.desktop.ui.tooling.preview.Preview
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Column
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.rounded.ArrowBack
-import androidx.compose.material.icons.rounded.ArrowBackIosNew
-import androidx.compose.material.icons.rounded.Settings
 import androidx.compose.material3.CenterAlignedTopAppBar
 import androidx.compose.material3.ExperimentalMaterial3Api
-import androidx.compose.material3.Icon
-import androidx.compose.material3.IconButton
 import androidx.compose.material3.LocalContentColor
 import androidx.compose.material3.Text
 import androidx.compose.material3.TopAppBarDefaults
@@ -18,7 +9,6 @@ import androidx.compose.runtime.Composable
 import androidx.compose.runtime.CompositionLocalProvider
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.text.font.FontWeight
-import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import io.timemates.app.style.system.theme.AppTheme
 
@@ -61,45 +51,4 @@ internal fun AppBarText(
         fontWeight = FontWeight.ExtraBold,
         color = AppTheme.colors.primary,
     )
-}
-
-@Preview
-@Composable
-private fun AppBarPreview() {
-    AppTheme {
-        Column(verticalArrangement = Arrangement.spacedBy(16.dp)) {
-            AppBar(
-                title = "TimeMates",
-            )
-
-            AppBar(
-                title = "TimeMates",
-                navigationIcon = {
-                    IconButton(
-                        onClick = {},
-                    ) {
-                        Icon(imageVector = Icons.Rounded.ArrowBack, contentDescription = null)
-                    }
-                }
-            )
-
-            AppBar(
-                title = "TimeMates",
-                navigationIcon = {
-                    IconButton(
-                        onClick = {},
-                    ) {
-                        Icon(imageVector = Icons.Rounded.ArrowBackIosNew, contentDescription = null)
-                    }
-                },
-                action = {
-                    IconButton(
-                        onClick = {},
-                    ) {
-                        Icon(imageVector = Icons.Rounded.Settings, contentDescription = null)
-                    }
-                }
-            )
-        }
-    }
 }
